separate line format definition from flat file entry
