+++ src/main/scala/flatFileEntry.scala
@@ -2,7 +2,6 @@ package bio4j.data.uniprot
 
 import java.time.LocalDate
 
-
 case class FlatFileEntry(
   val ID_line         : Line,
   val AC_lines        : Seq[Line], // nonEmpty
@@ -174,142 +173,12 @@ extends AnyEntry {
   lazy val sequence: Sequence = ???
 }
 
-
-sealed trait LineType { lazy val asString: String = toString }
-  case object ID extends LineType
-  case object AC extends LineType
-  case object DT extends LineType
-  case object DE extends LineType
-  case object GN extends LineType
-  case object OS extends LineType
-  case object OG extends LineType
-  case object OX extends LineType
-  // reference lines begin
-  case object RN extends LineType
-  case object RP extends LineType
-  case object RC extends LineType
-  case object RX extends LineType
-  case object RG extends LineType
-  case object RA extends LineType
-  case object RT extends LineType
-  case object RL extends LineType
-  // reference lines end
-  case object OH extends LineType
-  case object CC extends LineType
-  case object DR extends LineType
-  case object PE extends LineType
-  case object KW extends LineType
-  case object FT extends LineType
-  case object SQ extends LineType
-
-case object LineType {
-
-  def fromString(rep: String): Option[LineType] =
-    rep match {
-      case ID.asString  => Some(ID)
-      case AC.asString  => Some(AC)
-      case DT.asString  => Some(DT)
-      case DE.asString  => Some(DE)
-      case GN.asString  => Some(GN)
-      case OS.asString  => Some(OS)
-      case OG.asString  => Some(OG)
-      case OX.asString  => Some(OX)
-      case OH.asString  => Some(OH)
-      case CC.asString  => Some(CC)
-      case DR.asString  => Some(DR)
-      case PE.asString  => Some(PE)
-      case KW.asString  => Some(KW)
-      case FT.asString  => Some(FT)
-      case SQ.asString  => Some(SQ)
-      case _            => None
-    }
-}
-
-case class Line(
-  val lineType: LineType,
-  val content: String
-) {
-
-  def ofType(some: LineType): Option[Line] =
-    if(lineType == some) Some(this) else None
-
-  def isOfType(some: LineType): Boolean =
-    lineType == some
-}
-
-case object Line {
-
-  def from(line: String): Option[Line] = {
-
-    val lineTypeRep = line take 2
-
-    (LineType fromString lineTypeRep) map { lineType =>
-
-      val lineContent = line drop 5
-
-      Line(lineType, lineContent)
-    }
-  }
-
-  def isOfType(lt: LineType)(line: String): Boolean =
-    (line take 2) == lt.asString
-
-  def isReferenceLine(line: String): Boolean =
-    isOfType(RN)(line) ||
-    isOfType(RP)(line) ||
-    isOfType(RC)(line) ||
-    isOfType(RX)(line) ||
-    isOfType(RG)(line) ||
-    isOfType(RA)(line) ||
-    isOfType(RT)(line) ||
-    isOfType(RL)(line)
-
-
-  def contentOf(line: String): String =
-    line drop 5
-}
-
-case object parsers {
-
-  def entries(lines: Iterator[String]) = new Iterator[Seq[String]] {
-
-    private val rest: BufferedIterator[String] = lines.buffered
-
-    def hasNext: Boolean =
-      rest.hasNext
-
-    def next(): Seq[String] =
-      entry
-
-    @annotation.tailrec
-    private def entry_rec(acc: Array[String]): Array[String] =
-      if (rest.hasNext) {
-        if( rest.head.startsWith("//") ) {
-
-          val drop = rest.next()
-          acc
-        }
-        else entry_rec(acc :+ rest.next())
-      }
-      else acc
-
-    private def entry: Seq[String] = entry_rec(Array())
-  }
+case object FlatFileEntry {
 
   import Line._
-
-  // see http://stackoverflow.com/a/33521793/614394
-  lazy val localDateFormatter =
-    new java.time.format.DateTimeFormatterBuilder()
-      .parseCaseInsensitive()
-      .appendPattern("dd-MMM-yyyy")
-      .toFormatter(java.util.Locale.ENGLISH)
-
-  def localDateFrom(rep: String): LocalDate =
-    LocalDate.parse(rep, localDateFormatter)
-
+  
   // super ugly, but I don't see any simpler way
-  def flatFileEntryFrom(allLines: Seq[String]): FlatFileEntry = {
+  def from(allLines: Seq[String]): FlatFileEntry = {
 
     val (id_lines, rest0) = allLines.span(isOfType(ID))
     val (ac_lines, rest1) = rest0.span(isOfType(AC))
